Figure out how to use input fields with Reagent
This gives us a working example of numeric
input fields (with some limited error checking) for
use with things like population size and the number
of generations.

There is also a working example of a dropdown list
for use with things like parent selection, where we're
selecting from a fixed set of values.