Add inc/dec integer instructions
In watching this run on the symbolic regression
problems, the endgame is often a matter of adding
or subtracting a small number like 1 or 3. I thought
that adding the inc/dec instructions would potentially
help find a solution in those cases, and eyeball
statistics (dangerous I realize) suggest that it might?